fix(desktop): start bounded workers only for mounted diffs
