+++ products/desktop/docs/PERFORMANCE.md
@@ -1,0 +1,17 @@
+# Desktop performance
+
+## Diff workers
+
+Diff components own the lifetime of the syntax-highlighting worker pool.
+Opening a conversation without a diff does not start workers.
+Mounted diffs share one pool per renderer, capped at two workers and 200 entries in each AST cache.
+The last diff unmounting releases the pool.
+Languages load as files need them instead of preloading a review-wide language list.
+
+`DiffWorkerPool` supplies the same configuration to conversation previews and code review.
+Place new diff renderers inside this boundary, rather than adding a provider around a whole page.
+This keeps the worker budget independent of which view opens first.
+
+To check the lifecycle in the development app, open a plain conversation, open a diff, and then leave the diff.
+Inspect worker targets in Chromium DevTools: the plain conversation should have no diff workers, the diff should have two, and leaving all diffs should release them.
+Check syntax highlighting and large file scrolling in both conversation previews and code review.
