+++ products/desktop/packages/ui/src/features/sessions/components/ConversationView.tsx
@@ -1,6 +1,4 @@
 import { ArrowDown, XCircle } from "@phosphor-icons/react";
-import { WorkerPoolContextProvider } from "@pierre/diffs/react";
-import { useService } from "@posthog/di/react";
 import {
   Button,
   cn,
@@ -45,7 +43,6 @@ import {
   type VirtualizedListHandle,
 } from "@posthog/ui/features/sessions/components/VirtualizedList";
 import { CHAT_CONTENT_MAX_WIDTH } from "@posthog/ui/features/sessions/constants";
-import { DIFFS_HIGHLIGHTER_OPTIONS } from "@posthog/ui/features/sessions/diffHighlighterOptions";
 import { useConversationItems } from "@posthog/ui/features/sessions/hooks/useConversationItems";
 import { useConversationSearch } from "@posthog/ui/features/sessions/hooks/useConversationSearch";
 import {
@@ -63,10 +60,6 @@ import { SessionTaskIdProvider } from "@posthog/ui/features/sessions/useSessionT
 import { useSettingsStore } from "@posthog/ui/features/settings/settingsStore";
 import { TIP_KEYS } from "@posthog/ui/features/settings/tipKeys";
 import { SkillButtonActionMessage } from "@posthog/ui/features/skill-buttons/components/SkillButtonActionMessage";
-import {
-  DIFF_WORKER_FACTORY,
-  type DiffWorkerFactory,
-} from "@posthog/ui/shell/diffWorkerHost";
 import { Box, Flex, Text } from "@radix-ui/themes";
 import {
   memo,
@@ -117,19 +110,6 @@ export function ConversationView({
   scrollX = true,
   promptRecallRef,
 }: ConversationViewProps) {
-  const diffWorkerFactory = useService<DiffWorkerFactory>(DIFF_WORKER_FACTORY);
-  const diffsPoolOptions = useMemo(
-    () => ({
-      workerFactory: () => diffWorkerFactory(),
-      totalASTLRUCacheSize: 200,
-      // Each pooled highlighter worker is a full V8 isolate with shiki
-      // grammars loaded (~40MB RSS); the library default of 8 costs hundreds
-      // of MB for parallelism conversation diffs don't need.
-      poolSize: 2,
-    }),
-    [diffWorkerFactory],
-  );
-
   const listRef = useRef<VirtualizedListHandle>(null);
   const isAtBottomRef = useRef(true);
   const [showScrollButton, setShowScrollButton] = useState(false);
@@ -506,70 +486,65 @@ export function ConversationView({
   );
 
   return (
-    <WorkerPoolContextProvider
-      poolOptions={diffsPoolOptions}
-      highlighterOptions={DIFFS_HIGHLIGHTER_OPTIONS}
+    <div
+      ref={containerRef}
+      className="group/thread relative flex-1"
+      onPointerDownCapture={clearKeyboardFocus}
     >
-      <div
-        ref={containerRef}
-        className="group/thread relative flex-1"
-        onPointerDownCapture={clearKeyboardFocus}
-      >
-        {search.open && (
-          <ConversationSearchBar
-            ref={search.searchBarRef}
-            query={search.query}
-            currentMatch={search.currentIndex}
-            totalMatches={search.totalMatches}
-            onQueryChange={search.setQuery}
-            onNext={search.next}
-            onPrev={search.prev}
-            onClose={search.close}
-          />
-        )}
-
-        <MessageJumpPicker
-          open={jumpPickerOpen}
-          onOpenChange={setJumpPickerOpen}
-          items={items}
-          onJumpToMessage={handleJumpToMessage}
+      {search.open && (
+        <ConversationSearchBar
+          ref={search.searchBarRef}
+          query={search.query}
+          currentMatch={search.currentIndex}
+          totalMatches={search.totalMatches}
+          onQueryChange={search.setQuery}
+          onNext={search.next}
+          onPrev={search.prev}
+          onClose={search.close}
         />
+      )}
 
-        <SessionTaskIdProvider taskId={taskId}>
-          <VirtualizedList<ConversationTurn>
-            ref={listRef}
-            items={turns}
-            getItemKey={getTurnKey}
-            renderItem={renderTurn}
-            onScrollStateChange={handleScrollStateChange}
-            keepMounted={turnKeepMounted}
-            className="absolute inset-0 bg-background"
-            itemClassName="mx-auto px-2"
-            itemStyle={{ maxWidth: CHAT_CONTENT_MAX_WIDTH }}
-            footer={footer}
-            scrollX={scrollX}
-          />
-        </SessionTaskIdProvider>
-        {showScrollButton && (
-          <Box className="absolute right-6 bottom-4 z-10">
-            <Tooltip>
-              <TooltipTrigger
-                render={
-                  <Button
-                    size="icon-lg"
-                    variant="outline"
-                    onClick={scrollToBottom}
-                  >
-                    <ArrowDown size={14} weight="bold" />
-                  </Button>
-                }
-              />
-              <TooltipContent>Scroll to bottom</TooltipContent>
-            </Tooltip>
-          </Box>
-        )}
-      </div>
-    </WorkerPoolContextProvider>
+      <MessageJumpPicker
+        open={jumpPickerOpen}
+        onOpenChange={setJumpPickerOpen}
+        items={items}
+        onJumpToMessage={handleJumpToMessage}
+      />
+
+      <SessionTaskIdProvider taskId={taskId}>
+        <VirtualizedList<ConversationTurn>
+          ref={listRef}
+          items={turns}
+          getItemKey={getTurnKey}
+          renderItem={renderTurn}
+          onScrollStateChange={handleScrollStateChange}
+          keepMounted={turnKeepMounted}
+          className="absolute inset-0 bg-background"
+          itemClassName="mx-auto px-2"
+          itemStyle={{ maxWidth: CHAT_CONTENT_MAX_WIDTH }}
+          footer={footer}
+          scrollX={scrollX}
+        />
+      </SessionTaskIdProvider>
+      {showScrollButton && (
+        <Box className="absolute right-6 bottom-4 z-10">
+          <Tooltip>
+            <TooltipTrigger
+              render={
+                <Button
+                  size="icon-lg"
+                  variant="outline"
+                  onClick={scrollToBottom}
+                >
+                  <ArrowDown size={14} weight="bold" />
+                </Button>
+              }
+            />
+            <TooltipContent>Scroll to bottom</TooltipContent>
+          </Tooltip>
+        </Box>
+      )}
+    </div>
   );
 }
